+++ src/app/src/util/constants.js
@@ -226,6 +226,7 @@ export const dashboardClaimsRoute = '/dashboard/claims';
 export const claimedFacilitiesRoute = '/claimed';
 export const claimedFacilitiesDetailRoute = '/claimed/:claimID';
 export const dashboardClaimsDetailsRoute = '/dashboard/claims/:claimID';
+export const aboutClaimedFacilitiesRoute = '/about/claimedfacilities';
 
 export const contributeCSVTemplate =
     'country,name,address\nEgypt,Elite Merchandising Corp.,St. 8 El-Amrya Public Free Zone Alexandria Iskandariyah 23512 Egypt';